* origin/dev-cpp:
  Should fix all GCC warnings.
  Working compilation via GCC for TradeSignal.test.cpp and TradeSignalManager.test.cpp.
  State in the art of MQL/C++ compatibility.
  Making MQL code to compile in C++. Especially the TradeSignalManager.test.cpp.
  Trade: Fixes C++ compilation errors
  Adds TradeSignal and TradeSignalManager C++ compilation tests
  GHA: Compiles based on the file list
  Adds Compile workflow for CPP
  Adds StringSplit() as extern
  Adds .extern.h file for DateTime
  Refs: Uses PTR_ATTRIB syntax
  Integrates extern includes
  Adds Common and File defines
  Adds code page defines
  Adds .extern.h files
  Moves enums and structs from Std and MQL5 to appropriate files
  Moves Deal enums to Deal.enum.h

+++ SymbolInfo.struct.static.h
@@ -26,8 +26,9 @@
 #endif
 
 #include "MQL5.mqh"
-#include "Tick.struct.h"
+#include "Order.enum.h"
 #include "Std.h"
+#include "Tick.struct.h"
 
 /**
  * Struct to provide symbol information.
@@ -137,7 +138,8 @@ struct SymbolInfoStatic {
    * See: http://forum.mql4.com/30672
    */
   static unsigned int GetPointsPerPip(string _symbol) {
-    return (unsigned int)pow(10, SymbolInfoStatic::GetDigits(_symbol) - SymbolInfoStatic::GetPipDigits(_symbol));
+    return (unsigned int)pow((unsigned int)10,
+                             SymbolInfoStatic::GetDigits(_symbol) - SymbolInfoStatic::GetPipDigits(_symbol));
   }
 
   /**
@@ -263,7 +265,7 @@ struct SymbolInfoStatic {
    * Get real spread based on the ask and bid price (in points).
    */
   static unsigned int GetRealSpread(double _bid, double _ask, unsigned int _digits) {
-    return (unsigned int)round((_ask - _bid) * pow(10, _digits));
+    return (unsigned int)round((_ask - _bid) * pow((unsigned int)10, _digits));
   }
   static unsigned int GetRealSpread(string _symbol) {
     return GetRealSpread(SymbolInfoStatic::GetBid(_symbol), SymbolInfoStatic::GetAsk(_symbol),
